+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## Unreleased
 * Frontend: Add [linter based on eslint](https://angular.dev/cli/lint)
+* Frontend; Use [inject() instead of constructor injector](https://angular.dev/reference/migrations/inject-function)
 
 ## [0.0.11] - 2025-11-21
 * Backend: Update to [JDK 25 LTS](https://openjdk.org/projects/jdk/25/)